fix(tournaments): constrain rating to integers in Student schema

+++ src/lib/sheets/schemas.ts
@@ -27,7 +27,7 @@ export const StudentRow = z.object({
   last_payment_date: z.string().nullable().default(null),
   active: Bool,
   is_tournament_only: z.boolean().default(false),
-  rating: z.number().default(1000),
+  rating: z.number().int().default(1000),
   public_slug: z.string().nullable().default(null),
 });
 export type Student = z.infer<typeof StudentRow>;
